+++ _tasks/_consts.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+
+class Tasks(str, Enum):
+    test_task = "test-task"