+++ gcip2/__init__.py
@@ -1,5 +1,5 @@
 from gcip2.pipeline_core.gitlab_ci import GitlabCiBuilderImpl
-from gcip2.pipeline_core.jobs.base import Base, BaseLinux
+from gcip2.pipeline_core.jobs.base import Base, BaseLinux, BaseTask
 from gcip2.pipeline_core.pipeline import PipelineBuilderImpl
 from gcip2.pipeline_vars import Predefined
 from gcip2.vault import SecretsHandler
@@ -9,6 +9,7 @@ __all__ = (
     "GitlabCiBuilderImpl",
     "Base",
     "BaseLinux",
+    "BaseTask",
     "Predefined",
     "SecretsHandler",
 )