+++ gcip2/pipeline_core/jobs/base.py
@@ -27,3 +27,19 @@ class BaseLinux(JobBuilderImpl):
         ]
         self.model.before_script = [self._before_script_linux.script]
         return self
+
+
+class BaseTask(JobBuilderImpl):
+    _base = BaseLinux
+    _scipt_linux: assets.ScriptLinux = assets.ScriptLinux.load()
+    _after_script_linux: assets.AfterScriptLinux = assets.AfterScriptLinux.load()
+    _before_script_linux: assets.BeforeScriptLinux = assets.BeforeScriptLinux.load()
+
+    def apply(self: Self) -> Self:
+        self.model.after_script = [
+            self._after_script_linux.script,
+            "echo Do nothing.",
+        ]
+        self.model.script = [self._scipt_linux.script]
+        self.model.before_script = [self._before_script_linux.script]
+        return self