+++ pipelines/publish/helpers/create_version.py
@@ -3,6 +3,7 @@ import pathlib
 import requests
 
 from gcip2 import Predefined
+from gcip2.logging import logger as LOGGER
 from gcip2.vault import SecretsHandler
 
 
@@ -24,7 +25,7 @@ def create_version_tag(version: str):
 
     if not r.ok:
         raise RuntimeError(f"failed tag creating: {r.json()}")
-    print("tag creating successful", flush=True)
+    LOGGER.info("tag creating successful")
 
 
 def get_tag_from_version_file(version_file: pathlib.Path) -> str:
@@ -38,11 +39,11 @@ def main():
 
     version = get_tag_from_version_file(version_file=version_file)
 
-    print(f"creating tag for release: {version}", flush=True)
+    LOGGER.info(f"creating tag for release: {version}")
     if Predefined.CI_DEFAULT_BRANCH.must() == Predefined.CI_COMMIT_BRANCH.getenv():
         create_version_tag(version=version)
     else:
-        print("DRY-RUN: not release branch")
+        LOGGER.warning("DRY-RUN: not release branch")
 
 
 if __name__ == "__main__":