+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Changelog
 
+### [02.08.26 - 0.0.13]
+- added `TaskBuilder` and `TaskBuilderImpl` for job task executing
+- added `InteractivePythonAction` and `InteractiveShlex` actions
+- added `ProjectConfig` to actions that provide tools like SecretsHandler or Extra values
+- added `BaseTask` job template
+- added tasks testing pipeline
+- added `tasks_core` package with CLI for running tasks
+- added `loguru` for logging instead of `builtins.print`
+- added `extra.tasks` to `ProjectConfig` with `module` field - path to `_tasks` dir in project 
+
 ### [02.08.26 - 0.0.12]
 - added readme to package